% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ae_aeacn_ds_disctx.R
\name{check_ae_aeacn_ds_disctx}
\alias{check_ae_aeacn_ds_disctx}
\title{Check if, whenever a patient experiences an AE leading to permanent drug discontinuation,
they also have a DS record indicating this.}
\usage{
check_ae_aeacn_ds_disctx(AE, DS, preproc = identity, ...)
}
\arguments{
\item{AE}{Adverse Events SDTM dataset with variables USUBJID, AEDECOD, and at least one AEACN* variable}

\item{DS}{Disposition SDTM dataset with variables USUBJID, DSCAT, DSSCAT, DSDECOD}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
boolean value if check returns 0 obs, otherwise return subset dataframe.
}
\description{
Flags patients with AE leading to "永久停药"
but no corresponding "治疗结束"/"研究结束" record in DS.
}
\examples{

# FAIL: USUBJID 2 has "永久停药" in AE but no "结束" record in DS
AE <- data.frame(
  USUBJID = c("1", "2", "3"),
  AESEQ = c(1, 1, 1),
  AESTDTC = rep("2020-05-05", 3),
  AETERM = c("头痛", "心肌梗死", "寒战"),
  AEDECOD = c("头痛", "心肌梗死", "寒战"),
  AEACN = c("无", "永久停药", "无"),
  AEACN1 = c("", "", "永久停药"),
  stringsAsFactors = FALSE
)
DS <- data.frame(
  USUBJID = c("1", "3"),
  DSCAT = c("受试者处置", "受试者处置"),
  DSSCAT = c("治疗结束", "研究结束"),
  DSDECOD = c("完成", "不良事件"),
  stringsAsFactors = FALSE
)
check_ae_aeacn_ds_disctx(AE, DS)

# PASS: all patients with "永久停药" have corresponding DS "结束" record
DS2 <- data.frame(
  USUBJID = c("1", "2", "3"),
  DSCAT = c("受试者分布事件", "受试者分布事件", "受试者分布事件"),
  DSSCAT = c("治疗结束", "治疗结束", "研究结束"),
  DSDECOD = c("完成", "不良事件", "不良事件"),
  stringsAsFactors = FALSE
)
check_ae_aeacn_ds_disctx(AE, DS2)

}
\author{
1
}
