% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ae_aeout_aeendtc_dm_dthdtc.R
\name{check_ae_aeout_aeendtc_dm_dthdtc}
\alias{check_ae_aeout_aeendtc_dm_dthdtc}
\title{Check for AE outcome (AEOUT) of 'DEATH' with non-missing resolution
date that is not equal to the death date}
\usage{
check_ae_aeout_aeendtc_dm_dthdtc(AE, DM, preproc = identity, ...)
}
\arguments{
\item{AE}{Adverse Events SDTM dataset with variables USUBJID, AETERM,
AEDTHDTC, AEENDTC, AEOUT}

\item{DM}{SDTM dataset with variable DTHDTC}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check looks for AEs with outcome of 'DEATH' but AE
resolution date is not equal to AE death date.
Note that these datapoints are not collected the same way for all trials -
some trials leave AEENDTC missing if it was unresolved at death date. Confirm
within your team before querying this issue.
}
\examples{

AE <- data.frame(
 USUBJID = 1:10,
  DOMAIN = "AE",
  AESEQ = 101:110,
  AEENDTC = c(NA, "NA", rep("2015-03-12", 4), NA, "2020-01-01", NA, ""),
  AEOUT = c("", "", "", "жӯ»дәЎ", "з—Ҡж„Ҳ/жҒўеӨҚ", rep("жӯ»дәЎ", 5)),
  AETERM = 1:10,
  AESPID = "FORMNAME-R:13/L:13XXXX",
  stringsAsFactors = FALSE
)
DM <- data.frame(
  USUBJID = 1:10,
  DTHDTC = c(NA, "NA", rep("2015-03-12", 4), NA, NA, "2020-01-01", ""),
  stringsAsFactors = FALSE
)
check_ae_aeout_aeendtc_dm_dthdtc(AE, DM)
check_ae_aeout_aeendtc_dm_dthdtc(AE, DM, preproc = ql_derive_seq)

AE$AESPID <- NULL
check_ae_aeout_aeendtc_dm_dthdtc(AE, DM)

DM$DTHDTC <- NULL
AE$AEOUT <- NULL
check_ae_aeout_aeendtc_dm_dthdtc(AE, DM)
}
\author{
JH
}
