% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ae_aesdth_dthdtc.R
\name{check_ae_aesdth_dthdtc}
\alias{check_ae_aesdth_dthdtc}
\title{Check AEs with AEDTHDTC value but AESDTH not "Y"}
\usage{
check_ae_aesdth_dthdtc(AE, DM, preproc = identity, ...)
}
\arguments{
\item{AE}{Adverse Event SDTM dataset with variables USUBJID,
AESDTH, AEDECOD, AETERM, and AESTDTC}

\item{DM}{SDTM dataset with variable DTHDTC}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
Boolean value for whether the check passed or failed, with 'msg'
attribute if the check failed
}
\description{
This check looks for DM entries with an DTHDTC (death date)
value and  AESDTH of AE not equal to "Y"
}
\examples{

AE <- data.frame(
  USUBJID = c(1:7),
  AEDECOD = c(letters[1:5], "", NA),
  AETERM = letters[1:7],
  AESDTH = "是",
  AESTDTC = c(1:7),
  AESEQ = c(11:17),
  AESPID = "FORMNAME-R:5/L:5XXXX",
  stringsAsFactors = FALSE
)

DM<- data.frame(
  USUBJID = c(1:7),
  DTHDTC = "2020-01-02"
)
# expect pass
check_ae_aesdth_dthdtc(AE,DM)
check_ae_aesdth_dthdtc(AE,DM, preproc = ql_derive_seq)

# expect fail
AE1 <- AE
AE1$AESDTH[3] <- "否"
check_ae_aesdth_dthdtc(AE1, DM)
check_ae_aesdth_dthdtc(AE1, DM, preproc = ql_derive_seq)

# expect fail with AESDTH = NA
AE2 <- AE
AE2$AESDTH[4] <- NA
check_ae_aesdth_dthdtc(AE2, DM)
check_ae_aesdth_dthdtc(AE2, DM, preproc = ql_derive_seq)

# non-required variable missing
AE2$AESEQ <- NULL
check_ae_aesdth_dthdtc(AE2, DM)
check_ae_aesdth_dthdtc(AE2, preproc = ql_derive_seq)

# required variable missing
AE2$AESDTH <- NULL
check_ae_aesdth_dthdtc(AE2, DM)
check_ae_aesdth_dthdtc(AE2, preproc = ql_derive_seq)
DM$DTHDTC <- NULL
check_ae_aesdth_dthdtc(AE1, DM)
check_ae_aesdth_dthdtc(AE1, DM, preproc = ql_derive_seq)
}
\author{
1
}
