% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ae_aestdtc_after_dd.R
\name{check_ae_aestdtc_after_dd}
\alias{check_ae_aestdtc_after_dd}
\title{Check for AE dates occurring after death date}
\usage{
check_ae_aestdtc_after_dd(AE, DM, preproc = identity, ...)
}
\arguments{
\item{AE}{Adverse Event SDTM dataset with variables USUBJID,
AESTDTC, AEDECOD, AETERM, AEGRPID (optional), AESPID (optional)}

\item{DM}{Demographics SDTM dataset with variables USUBJID,
DTHDTC}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
Boolean value for whether the check passed or failed, with 'msg'
attribute if the check failed
}
\description{
This check looks for AE dates that occur after death date
}
\examples{

AE <- data.frame(
  STUDYID = 1:5, USUBJID = LETTERS[1:5],
  AESTDTC = rep("2016-01-01", 5),
  AEDECOD = c("", "", rep("Myocarditis", 3)),
  AETERM = c("INJURY", rep("MYOCARDITIS", 4)),
  AESEQ = 1:5,
  AEGRPID = "01",
  AESPID = "01",
  stringsAsFactors = FALSE
)

DM <- data.frame(
  STUDYID = 1:5, USUBJID = LETTERS[1:5],
  DTHDTC = c(rep("", 4), "2016-01-02"),
  stringsAsFactors = FALSE
)

check_ae_aestdtc_after_dd(AE, DM)

AE$AESTDTC[1] <- "2016-01-03"
AE$USUBJID[1] <- AE$USUBJID[5]

check_ae_aestdtc_after_dd(AE, DM, preproc = ql_derive_seq)

AE$AESPID <- NULL
check_ae_aestdtc_after_dd(AE, DM)

AE$AESTDTC <- NULL
check_ae_aestdtc_after_dd(AE, DM)

}
\author{
1
}
