% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ae_fatal.R
\name{check_ae_fatal}
\alias{check_ae_fatal}
\title{Check for death variable consistency when AEOUT=="DEATH"}
\usage{
check_ae_fatal(AE, DM, preproc = identity, ...)
}
\arguments{
\item{AE}{Adverse Events SDTM dataset with variables USUBJID, AEDECOD,
AESTDTC, AEDTHDTC, AEOUT, AESDTH}

\item{DM}{SDTM dataset with variable DTHDTC}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check looks for consistency in AESDTH, AEDTHDTC, and
AETOXGR (if applicable) when AEOUT is 'DEATH'.  Note, this check expects
AE grade/severity variables to be populated for either all records or none.
In a case where both AETOXGR and AESEV exist and some records are supposed
to have AETOXGR populated while others have AESEV (ie the two variables are
mutually exclusive) then this check will likely return false positives.
}
\examples{

# AETOXGR, no AESEV

AE <- data.frame(
  USUBJID = 1:5,
  AESTDTC = "01JAN2017",
  AEDECOD = c("AE1", "AE2", "AE3", "AE4", "AE5"),
  AEOUT = "死亡",
  AESDTH = c("是", "是", "否", "是", NA),
  AETOXGR = c("5", "5", "5", NA, NA),
  AESEQ = 11:15,
  stringsAsFactors = FALSE
)

DM <- data.frame(
  USUBJID = 1:5,
  DTHDTC = c("01FEB2017", NA, "02FEB2017", "03FEB2017", NA),
  stringsAsFactors = FALSE
)

check_ae_fatal(AE, DM)
check_ae_fatal(AE, DM, preproc = ql_derive_seq)

AE$AETOXGR <- NULL
check_ae_fatal(AE, DM)

AE$AEDECOD <- NULL
check_ae_fatal(AE, DM)


# AESEV, no AETOXGR

AE <- data.frame(
  USUBJID = 1:5,
  AESTDTC = "01JAN2017",
  AEDECOD = c("AE1", "AE2", "AE3", "AE4", "AE5"),
  AEOUT = "死亡",
  AESDTH = c("是", "是", "否", "是", NA),
  AESEV = c("SEVERE", "MILD", "SEVERE", NA, NA),
  AESEQ = 11:15,
  stringsAsFactors = FALSE
)
DM <- data.frame(
  USUBJID = 1:5,
  DTHDTC = c("01FEB2017", "02FEB2017", "03FEB2017", "04FEB2017", NA),
  stringsAsFactors = FALSE
)

check_ae_fatal(AE, DM)
check_ae_fatal(AE, DM, preproc = ql_derive_seq)

AE$AESEV <- NULL
check_ae_fatal(AE, DM)

# Both AESEV and AETOXGR have non-missing values

AE <- data.frame(
  USUBJID = 1:7,
  AESTDTC = "01JAN2017",
  AEDECOD = c("AE1", "AE2", "AE3", "AE4", "AE5", "AE6", "AE7"),
  AEOUT = "死亡",
  AESDTH = c("是", "是", "否", "是", NA, "是", "是"),
  AESEV = c("SEVERE", "MILD", "SEVERE", NA, NA, "MILD", "SEVERE"),
  AETOXGR = c("5", "5", "5", NA, NA, "1", "5"),
  AESEQ = 11:17,
  stringsAsFactors = FALSE
)
DM <- data.frame(
  USUBJID = 1:7,
  DTHDTC = c("01FEB2017", NA, "02FEB2017", "03FEB2017", NA, "04FEB2017", "05FEB2017"),
  stringsAsFactors = FALSE
)

check_ae_fatal(AE, DM)
check_ae_fatal(AE, DM, preproc = ql_derive_seq)


# Neither AESEV or AETOXGR

AE <- data.frame(
  USUBJID = 1:5,
  AESTDTC = "01JAN2017",
  AEDECOD = c("AE1", "AE2", "AE3", "AE4", "AE5"),
  AEOUT = "死亡",
  AESDTH = c("是", "是", "否", "是", NA),
  AESPID = "FORMNAME-R:12/L:2XXXX",
  stringsAsFactors = FALSE
)
DM <- data.frame(
  USUBJID = 1:5,
  DTHDTC = c("01FEB2017", NA, "02FEB2017", "03FEB2017", NA),
  stringsAsFactors = FALSE
)
check_ae_fatal(AE, DM)

# AETOXGR exists but unmapped AESEV

AE <- data.frame(
  USUBJID = 1:5,
  AESTDTC = "01JAN2017",
  AEDECOD = c("AE1", "AE2", "AE3", "AE4", "AE5"),
  AEOUT = "死亡",
  AESDTH = c("是", "是", "否", "是", NA),
  AESEV = rep(NA, 5),
  AETOXGR = c("5", "5", "5", NA, NA),
  AESEQ = 111:115,
  stringsAsFactors = FALSE
)
DM <- data.frame(
  USUBJID = 1:5,
  DTHDTC = c("01FEB2017", NA, "02FEB2017", "03FEB2017", NA),
  stringsAsFactors = FALSE
)


check_ae_fatal(AE, DM)
check_ae_fatal(AE, DM, preproc = ql_derive_seq)

# AETOXGR and AESEV exist, by both are unmapped

AE <- data.frame(
  USUBJID = 1:5,
  AESTDTC = "01JAN2017",
  AEDECOD = c("AE1", "AE2", "AE3", "AE4", "AE5"),
  AEOUT = "死亡",
  AESDTH = c("是", "是", "否", "是", NA),
  AESEV = NA,
  AETOXGR = NA,
  AESEQ = "FORMNAME-R:12/L:2XXXX",
  stringsAsFactors = FALSE
)
DM <- data.frame(
  USUBJID = 1:5,
  DTHDTC = c("01FEB2017", NA, "02FEB2017", "03FEB2017", NA),
  stringsAsFactors = FALSE
)

check_ae_fatal(AE, DM)
check_ae_fatal(AE, DM, preproc = ql_derive_seq)
}
\author{
JH
}
