% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ae_withdr_ds_discon.R
\name{check_ae_withdr_ds_discon}
\alias{check_ae_withdr_ds_discon}
\title{Check if an AE leading to drug being withdrawn is reflected in DS}
\usage{
check_ae_withdr_ds_discon(AE, DS, TS, preproc = identity, ...)
}
\arguments{
\item{AE}{Adverse Events SDTM dataset with variables USUBJID, AEACN}

\item{DS}{Disposition SDTM dataset with variables USUBJID, DSCAT, DSSCAT}

\item{TS}{Trial Summary SDTM dataset with variables TSPARMCD, TSVAL}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This checks that if there is an AE with AEACN="DRUG WITHDRAWN" then there should be a treatment discontinuation
record indicated by DS.DSSCAT
}
\examples{

AE <- data.frame(
  USUBJID = c("001", "002"),
  AEACN = c("永久停药", "永久停药"),
  AEDECOD = c("恶心", "头痛"),
  stringsAsFactors = FALSE
)

DS <- data.frame(
  USUBJID = c("001", "003"),
  DSSCAT = c("研究中止", "研究结束"),
  DSCAT = c("处置事件", "处置事件"),
  DSDECOD = c("不良事件", "完成"),
  stringsAsFactors = FALSE
)

TS <- data.frame(
  TSPARMCD = c("TRT"),
  TSVAL = c("Drug A"),
  stringsAsFactors = FALSE
)
check_ae_withdr_ds_discon(AE, DS, TS)
}
\author{
1
}
