% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dd_ae_aeout_dthdtc.R
\name{check_dd_ae_aeout_dthdtc}
\alias{check_dd_ae_aeout_dthdtc}
\title{Check if there is a death date and AEOUT is death agreement}
\usage{
check_dd_ae_aeout_dthdtc(AE, DM, preproc = identity, ...)
}
\arguments{
\item{AE}{Adverse Events SDTM dataset with variables USUBJID, AEDTHDTC,
AEDECOD, AESTDTC and AEOUT}

\item{DM}{SDTM dataset with variable DTHDTC}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check looks for AE death dates if AEOUT is death and for the
reverse, i.e if there is an AE death date, then AEOUT should have the value
"死亡".
}
\examples{

AE <- data.frame(
  USUBJID = 1:3,
  AEDECOD = 1:3,
  AESTDTC = 1:3,
  AEOUT = rep("死亡", 3),
  AESEQ = 11:13,
  stringsAsFactors = FALSE
)
DM <- data.frame(
  USUBJID = 1:3,
  DTHDTC = c("2020-01-01", "2020-01-02", "2020-01-03"),
  stringsAsFactors = FALSE
)
# pass
check_dd_ae_aeout_dthdtc(AE, DM)

# fail - 1 case (AEDTHDTC not populated but AEOUT == FATAL)
AE1 <- AE
DM[3, "DTHDTC"] <- NA
check_dd_ae_aeout_dthdtc(AE1, DM)
check_dd_ae_aeout_dthdtc(AE1, DM, preproc = ql_derive_seq)

# fail - 2 case -- even though DTHDTC populated
AE2 <- AE
AE2[1, "AEOUT"] <- NA
check_dd_ae_aeout_dthdtc(AE2, DM)
check_dd_ae_aeout_dthdtc(AE2, DM, preproc = ql_derive_seq)

# 2 cases

AE[1, "AEOUT"] <- NA
check_dd_ae_aeout_dthdtc(AE, DM)
check_dd_ae_aeout_dthdtc(AE, DM, preproc = ql_derive_seq)

# 2 cases
AE[1, "AEOUT"] <- "NOT RECOVERED/NOT RESOLVED"
check_dd_ae_aeout_dthdtc(AE, DM)
check_dd_ae_aeout_dthdtc(AE, DM, preproc = ql_derive_seq)


# critical variables are missing
DM$DTHDTC <- NULL
AE$USUBJID <- NULL
check_dd_ae_aeout_dthdtc(AE, DM)
check_dd_ae_aeout_dthdtc(AE, DM, preproc = ql_derive_seq)

}
\author{
JH
}
