% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dd_dm_dthdtc_ds_dsstdtc.R
\name{check_dd_dm_dthdtc_ds_dsstdtc}
\alias{check_dd_dm_dthdtc_ds_dsstdtc}
\title{Check if death date is the same in DM and DS domains}
\usage{
check_dd_dm_dthdtc_ds_dsstdtc(DM, DS, preproc = identity, ...)
}
\arguments{
\item{DM}{SDTM dataset with variables USUBJID and DTHDTC}

\item{DS}{Disposition SDTM dataset with variables USUBJID, DSDECOD, DSSTDTC}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check compares death date in DM DTHDTC with death date in
DS DSSTDTC. It is expected that they are the same.
}
\examples{

DM <- data.frame(
  STUDYID = rep(1, 3),
  USUBJID = 1:3,
  DTHDTC = c("2020-01-01", "2020-01-02", "2020-01-03")
)

DS <- data.frame(
  STUDYID = rep(1, 3),
  USUBJID = 1:3,
  DSDECOD = rep("死亡", 3),
  DSSTDTC = c("2020-01-01", "2020-01-02", "2020-01-03"),
  DSSEQ = 11:13,
  stringsAsFactors = FALSE
)

# no case
check_dd_dm_dthdtc_ds_dsstdtc(DM, DS)

# 1 case
DS[3, "DSSTDTC"] <- "2000-01-01"
check_dd_dm_dthdtc_ds_dsstdtc(DM, DS, preproc = ql_derive_seq)

# check for non existence of vars
DS$DSDECOD <- NULL
DS$DSSTDTC <- NULL
check_dd_dm_dthdtc_ds_dsstdtc(DM, DS)

}
\author{
JH
}
