% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dm_ds_partial_death_dates.R
\name{check_dm_ds_partial_death_dates}
\alias{check_dm_ds_partial_death_dates}
\title{Check for partial death dates in DM and DS}
\usage{
check_dm_ds_partial_death_dates(DM, DS, preproc = identity, ...)
}
\arguments{
\item{DM}{SDTM dataset with variable DTHDTC}

\item{DS}{Dispostion SDTM dataset with variables USUBJID,DSSCAT,DSSTDTC,DSDECOD}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the test failed
}
\description{
This checks looks for partial death dates in DM and DS
}
\examples{

DM <- data.frame(
  USUBJID = 1:3,
  DTHDTC = c("2017-01-01", "2017", NA),
  stringsAsFactors = FALSE
)

DS <- data.frame(
  USUBJID = 1:4,
  DSSEQ = 11:14,
  DSSCAT = "STUDY DISCON",
  DSDECOD = "死亡",
  DSSTDTC = c("2017-01-01", "2017", "2017-01-02", "2016-10"),
  stringsAsFactors = FALSE
)

check_dm_ds_partial_death_dates(DM, DS)
check_dm_ds_partial_death_dates(DM, DS, preproc = ql_derive_seq)

DS$DSSTDTC <- NULL

check_dm_ds_partial_death_dates(DM, DS)
}
\author{
Will Harris
}
