% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dm_dthdtc_ds_death.R
\name{check_dm_dthdtc_ds_death}
\alias{check_dm_dthdtc_ds_death}
\title{Check for missing DTHDTC where DS indicates death}
\usage{
check_dm_dthdtc_ds_death(DM, DS)
}
\arguments{
\item{DM}{SDTM dataset with variables USUBJID, DTHDTC}

\item{DS}{Disposition SDTM dataset with variables USUBJID, DSDECOD, DSTERM, DSSTDTC}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check looks for missing DTHDTC values if a patient has a
DS record where DSDECOD="死亡"
}
\examples{

DM <- data.frame(
  USUBJID = 1:3,
  DTHDTC = c(NA, NA, "2020-01-01")
)


# older mapping
DS <- data.frame(
  USUBJID = 1:4,
  DSTERM = c(
    "死亡", "死亡",
    "死亡", "死亡"
  ),
  DSDECOD = rep("死亡", 4),
  DSSTDTC = "2020-01-01"
)

check_dm_dthdtc_ds_death(DM, DS)

DS$DSSTDTC <- NULL

check_dm_dthdtc_ds_death(DM, DS)

# newer mapping that
DS <- data.frame(
  USUBJID = 1:4,
  DSTERM = c(
    "不良事件", "不良事件",
    "不良事件", "死亡"
  ),
  DSDECOD = rep("死亡", 4),
  DSSTDTC = "2020-01-01"
)

# pass for study with newer mapping, as another function (check_dd_death_date.R) covers this
check_dm_dthdtc_ds_death(DM, DS)
}
\author{
Jh
}
