% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ds_dsdecod_dsstdtc.R
\name{check_ds_dsdecod_dsstdtc}
\alias{check_ds_dsdecod_dsstdtc}
\title{Check DS with death record but no death date}
\usage{
check_ds_dsdecod_dsstdtc(DS)
}
\arguments{
\item{DS}{Disposition SDTMv dataset with variables USUBJID, DSDECOD,
DSSCAT and DSSTDTC}
}
\value{
Boolean value for whether the check passed or failed, with 'msg'
attribute if the test failed
}
\description{
This check looks for patients in DS who have a record indicating
death but no corresponding record with death date in DS.
For example, "生存随访" records often have no death dates, so for
a data cut to be applied properly, you have to impute that missing
death date from another record where its not missing (e.g. Study Discon form)
}
\examples{

DS1 <- data.frame(
  STUDYID = rep(1, 4),
  USUBJID = c(1, 1, 2, 3),
  DSDECOD = c("死亡", "死亡相关", "完成", "不良事件"),
  DSSCAT = c("处置事件", "处置事件", "方案里程碑", "受试者分布事件"),
  DSSTDTC = c(NA, "2016-01-01", "2016-01-03", "2016-01-02"),
  stringsAsFactors = FALSE
)

check_ds_dsdecod_dsstdtc(DS1)

DS2 <- DS1
DS2$DSSTDTC[2] <- NA

check_ds_dsdecod_dsstdtc(DS2)

}
\author{
1
}
