% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ec_dup.R
\name{check_ec_dup}
\alias{check_ec_dup}
\title{Check for duplicate EC records}
\usage{
check_ec_dup(EC)
}
\arguments{
\item{EC}{EC SDTM dataset with variables USUBJID, ECTRT, ECDOSE, ECSTDTC.
VISIT is optional.}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check looks for duplicate treatment records in EC
}
\examples{

# FAIL: duplicate EC records
EC <- data.frame(
  USUBJID = rep(1, 2),
  ECTRT = rep("药物A", 2),
  ECDOSE = rep(10, 2),
  ECSTDTC = rep("2020-01-01", 2),
  ECOCCUR = "是",
  stringsAsFactors = FALSE
)
check_ec_dup(EC)

# PASS: no duplicates
EC2 <- data.frame(
  USUBJID = 1:2,
  ECTRT = c("药物A", "药物B"),
  ECDOSE = c(10, 20),
  ECSTDTC = c("2020-01-01", "2020-01-02"),
  ECOCCUR = "是",
  stringsAsFactors = FALSE
)
check_ec_dup(EC2)

# FAIL: duplicates created by rbind
EC3 <- rbind(EC2, EC2)
check_ec_dup(EC3)

# missing required variable
EC4 <- EC2
EC4$ECTRT <- NULL
check_ec_dup(EC4)

}
\author{
1
}
