% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ec_ecdose_pos_ecoccur_no.R
\name{check_ec_ecdose_pos_ecoccur_no}
\alias{check_ec_ecdose_pos_ecoccur_no}
\title{check ec exdose pos exoccur_no}
\usage{
check_ec_ecdose_pos_ecoccur_no(EC, drug = NULL)
}
\arguments{
\item{EC}{SDTM dataset}

\item{drug}{TRT A}
}
\value{
Returns message depending on if there are records with positive dose
}
\description{
EX Valid Dose, Missing EXOCCUR	Records with missing drug administration
occurrence  info but non-missing dose value
}
\examples{
EC <- data.frame(
  USUBJID = 1:6,
  ECSTDTC = rep("2017-01-01", 6),
  ECTRT = c(rep("TRT A", 3), rep("TRT B", 3)),
  ECOCCUR = c("是", "否", ".", "", "是", "否"),
  ECDOSE = c(0, 10, 5, 15, 0, 8),
  VISIT = "VISIT 1",
  ECMOOD = "已执行",
  stringsAsFactors = FALSE
)

check_ec_ecdose_pos_ecoccur_no(EC)

check_ec_ecdose_pos_ecoccur_no(EC, drug = "TRT A")
check_ec_ecdose_pos_ecoccur_no(EC, drug = "TRT B")

EC$ECDOSE <- NULL

check_ec_ecdose_pos_ecoccur_no(EC)

}
