% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ec_ecdosu.R
\name{check_ec_ecdosu}
\alias{check_ec_ecdosu}
\title{Check for missing ECDOSU records}
\usage{
check_ec_ecdosu(EC)
}
\arguments{
\item{EC}{SDTM dataset with variables USUBJID,ECTRT,ECSTDTC,ECDOSU}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check looks for missing ECODOSU values for valid doses(ECMOOD="是" and ECOCCUR = "是")
}
\examples{

EC <- data.frame(
  USUBJID = 1:10,
  ECTRT = 1:10,
  ECSTDTC = 1:10,
  ECDOSE = 1:10,
  ECOCCUR = as.character(c(rep("是", 5), rep("否", 5))),
  ECDOSU = as.character(rep("mg", 10)),
  ECMOOD = "已执行"
)

EC$ECDOSU[1] <- ""
EC$ECDOSU[2] <- "NA"
EC$ECDOSU[3] <- NA

check_ec_ecdosu(EC)

EC$ECSTDTC <- NULL

check_ec_ecdosu(EC)

}
\author{
Jen Chen
}
