% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ec_ecoccur_mis_ecdose_nonmis.R
\name{check_ec_ecoccur_mis_ecdose_nonmis}
\alias{check_ec_ecoccur_mis_ecdose_nonmis}
\title{EC Missing Occurrence but Non-Missing Dose Check}
\usage{
check_ec_ecoccur_mis_ecdose_nonmis(EC)
}
\arguments{
\item{EC}{EC 域数据框，包含 USUBJID, ECTRT, ECOCCUR, ECDOSE, ECSTDTC, ECMOOD}
}
\value{
带属性的布尔值，存在问题返回 FALSE 并附带数据
}
\description{
Records with missing drug administration occurrence info but non-missing dose value
}
\examples{

EC <- data.frame(
  USUBJID = 1:5,
  ECSTDTC = rep("2017-01-01", 5),
  ECTRT = c(rep("TRT A", 2), rep("TRT B", 3)),
  ECOCCUR = c("是", NA, ".", "", "否"),
  ECDOSE = c(10, 15, 0, 20, 5),
  ECMOOD = "已执行",
  stringsAsFactors = FALSE
)

check_ec_ecoccur_mis_ecdose_nonmis(EC)

EC2 <- data.frame(
  USUBJID = 1:4,
  ECSTDTC = rep("2017-01-01", 4),
  ECTRT = rep("TRT A", 4),
  ECOCCUR = c("是", "否", "是", "否"),
  ECDOSE = c(10, 15, 20, 5),
  ECMOOD = "已执行",
  stringsAsFactors = FALSE
)

check_ec_ecoccur_mis_ecdose_nonmis(EC2)

EC3 <- EC
EC3$ECOCCUR <- NULL

check_ec_ecoccur_mis_ecdose_nonmis(EC3)

}
