% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ex_exstdtc_after_dd.R
\name{check_ex_exstdtc_after_dd}
\alias{check_ex_exstdtc_after_dd}
\title{Check for EX dates occurring after death date}
\usage{
check_ex_exstdtc_after_dd(DM, EX)
}
\arguments{
\item{DM}{Demographics SDTM dataset with variables USUBJID,
DTHDTC}

\item{EX}{Exposure SDTM dataset with variables USUBJID,
EXSTDTC, EXTRT, and EXDOSE. EXOCCUR is optional and, when available,
only EXOCCUR = "是" records are checked.}
}
\value{
Boolean value for whether the check passed or failed, with 'msg'
attribute if the check failed
}
\description{
This check looks for EX dates that occur after death date
}
\examples{

DM <- data.frame(
  STUDYID = 1:5, USUBJID = LETTERS[1:5],
  DTHDTC = c(rep("", 4), "2016-01-02"),
  stringsAsFactors = FALSE
)

EX <- data.frame(
  STUDYID = 1:5, USUBJID = LETTERS[1:5],
  EXSTDTC = rep("2015-12-31", 5),
  EXTRT = LETTERS[1:5],
  EXDOSE = 1:5,
  stringsAsFactors = FALSE
)

check_ex_exstdtc_after_dd(DM, EX)

EX$EXSTDTC[1] <- "2016-01-03"
EX$USUBJID[1] <- EX$USUBJID[5]

check_ex_exstdtc_after_dd(DM, EX)

}
\author{
1
}
