% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lb_lbdtc_visit_ordinal_error.R
\name{check_lb_lbdtc_visit_ordinal_error}
\alias{check_lb_lbdtc_visit_ordinal_error}
\title{Check that all LB dates are duplicated or earlier than last
visit's (possible datetime data entry error)}
\usage{
check_lb_lbdtc_visit_ordinal_error(LB)
}
\arguments{
\item{LB}{SDTM dataset with variables USUBJID, VISITNUM, VISIT, LBDTC, LBSTAT.
LBTESTCD/LBCAT are used for grouping when available.}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check identifies LBDTC values that are duplicated or
earlier than last visit's. Records with LBSTAT == '未查' and unscheduled
visits (VISIT with the string "计划外") are excluded.
}
\examples{

# no case
LB1 <- data.frame(
  USUBJID = c(rep("101", 5), rep("102", 5)),
  LBCAT = "Hematology",
  LBDTC = rep(c(
    "2017-01-01T08:25",
    "2017-01-05T09:25",
    "2017-01-15T10:25",
    "2017-01-20T08:25",
    "2017-01-25T08:25"
  ), 2),
  VISITNUM = rep(1:5, 2),
  VISIT = rep(c(
    "C1/D1",
    "C2/D1",
    "C3/D1",
    "计划外访视",
    "C4/D1"
  ), 2),
  LBSTAT = c(rep("", 9), "未查"),
  stringsAsFactors = FALSE
)

check_lb_lbdtc_visit_ordinal_error(LB1)

LB2 <- LB1
LB2$LBCAT <- "病毒学检查"
LB3 <- rbind(LB1, LB2)
check_lb_lbdtc_visit_ordinal_error(LB3)

# adding cases with earlier date
LB3$LBDTC[LB3$USUBJID == 101 & LB3$VISIT == "C3/D1"] <- "2016-01-10T08:25"
LB3$LBDTC[LB3$USUBJID == 102 & LB3$VISIT == "C2/D1"] <- "2016-01-01T06:25"
check_lb_lbdtc_visit_ordinal_error(LB = LB3)

# adding cases with duplicated date
LB3$LBDTC[LB3$USUBJID == 102 & LB3$VISIT == "C3/D1"] <- "2017-01-15T10:25"
LB3 <- LB3[order(LB3$USUBJID, LB3$VISITNUM, LB3$LBDTC), ]
check_lb_lbdtc_visit_ordinal_error(LB = LB3)

# check if all NOT DONE
LB4 <- LB3
LB4$LBSTAT <- "未查"
check_lb_lbdtc_visit_ordinal_error(LB = LB4)

# check dropping a required variable
LB4$LBSTAT <- NULL
check_lb_lbdtc_visit_ordinal_error(LB = LB4)

}
\author{
1
}
