% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_qs_dup.R
\name{check_qs_dup}
\alias{check_qs_dup}
\title{Check for multiple dates at the same visit in QS}
\usage{
check_qs_dup(QS)
}
\arguments{
\item{QS}{QS SDTM dataset with variables USUBJID, QSCAT, QSTESTCD, VISIT, QSDTC}
}
\value{
Boolean value for whether the check passed or failed, with 'msg'
attribute if the test failed
}
\description{
Identifies multiple dates at the same visit in QS by
USUBJID, QSCAT, QSTESTCD, and VISIT.
Unscheduled visits are excluded.
}
\examples{

QS1 <- data.frame(
  USUBJID = c(rep(101, 5), rep(102, 5)),
  QSCAT = "DLQI",
  QSTESTCD = "DLQI01",
  QSDTC = rep(c(
    "2017-01-01T08:25", "2017-01-05T09:25",
    "2017-01-15T10:25", "2017-01-20T08:25", "2017-01-25T08:25"
  ), 2),
  VISITNUM = rep(1:5, 2),
  VISIT = rep(c("C1/D1", "C2/D1", "C3/D1", "计划外", "C4/D1"), 2),
  stringsAsFactors = FALSE
)
check_qs_dup(QS = QS1)

# multiple dates for the same visit in QS
QS2 <- QS1
QS2$VISIT[QS2$USUBJID == 101] <- "C1/D1"
check_qs_dup(QS = QS2)

}
\author{
1
}
