% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_qs_qsdtc_after_dd.R
\name{check_qs_qsdtc_after_dd}
\alias{check_qs_qsdtc_after_dd}
\title{Check for QS dates occurring after death date}
\usage{
check_qs_qsdtc_after_dd(DM, QS)
}
\arguments{
\item{DM}{Demographics SDTM dataset with variables USUBJID,
DTHDTC}

\item{QS}{Questionnaire Test Findings SDTM dataset with variables
USUBJID, QSDTC, QSCAT, and QSORRES}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the test failed
}
\description{
This check looks for QS dates that occur after death date
}
\examples{

DM <- data.frame(
  USUBJID = c(1, 2),
  DTHDTC = c("2016-01-01", "2016-01-01"),
  stringsAsFactors = FALSE
)

QS <- data.frame(
  USUBJID = c(1, 1, 1, 2, 2, 2),
  QSDTC = c(
    "2015-06-30", "2015-09-30", "2015-12-30",
    "2015-06-30", "2015-09-30", "2015-12-30"
  ),
  QSCAT = "A",
  QSORRES = LETTERS[1:6],
  QSSTAT = "",
  VISIT = c("Week 1", "Week 12", "Week 24", "Week 1", "Week 12", "Week 24"),
  QSSTRESC = LETTERS[1:6],
  stringsAsFactors = FALSE
)

check_qs_qsdtc_after_dd(DM, QS)

QS$QSDTC[3:5] <- "2016-01-03"
check_qs_qsdtc_after_dd(DM, QS)

QS$QSSTAT[3] <- "Not Done"
check_qs_qsdtc_after_dd(DM, QS)

DM1 <- data.frame(
  USUBJID = 1,
  DTHDTC = "2016-01",
  stringsAsFactors = FALSE
)

QS1 <- data.frame(
  USUBJID = 1,
  QSDTC = c("2015-06-30", "2016-01-15", "2016-01-15"),
  QSCAT = rep("EQ-5D-5L"),
  QSORRES = "1",
  QSSTAT = "",
  VISIT = c("Week 1", "Week 12", "Week 12"),
  QSSTRESC = "1",
  stringsAsFactors = FALSE
)

check_qs_qsdtc_after_dd(DM = DM1, QS = QS1)
}
\author{
1
}
