% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rs_rsdtc_across_visit.R
\name{check_rs_rsdtc_across_visit}
\alias{check_rs_rsdtc_across_visit}
\title{Check RS records where the same date occurs across multiple visits}
\usage{
check_rs_rsdtc_across_visit(RS, preproc = identity, ...)
}
\arguments{
\item{RS}{Disease Response SDTM dataset with variables USUBJID, RSDTC, VISIT,
RSEVAL (optional), RSTESTCD (optional), RSLNKGRP (optional), RSSPID (optional)}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check identifies records where the same date RSDTC occurs
across multiple visits. Only applies to assessments by investigator,
selected based on uppercased RSEVAL = "研究者" or missing or
RSEVAL variable does not exist.
}
\examples{

# example that will be flagged
RS <- data.frame(
  USUBJID = 1,
  RSDTC = c(rep("2016-01-01", 3), rep("2016-06-01", 5), rep("2016-06-24", 2)),
  VISIT = c(rep("C1D1", 3), rep("C1D2", 3), rep("C2D1", 4)),
  RSSEQ = 1:10,
  RSSPID = c(sprintf("\%02d", 1:6), "01", sprintf("\%02d", 1:3)),
  RSEVAL = c(rep("研究者", 10)),
  RSTESTCD = c(rep("OVRLRESP")),
  stringsAsFactors = FALSE
)

check_rs_rsdtc_across_visit(RS)
check_rs_rsdtc_across_visit(RS, preproc = ql_derive_seq)

# example that will not be flagged because not Investigator
RS0 <- RS
RS0$RSEVAL <- "独立评估者"
check_rs_rsdtc_across_visit(RS0)
check_rs_rsdtc_across_visit(RS0, preproc = ql_derive_seq)

# example with log line differences in Rave form with records flagged
RS1 <- RS
RS1$RSSPID <- c(
  rep("01", 4),
  rep("02", 2),
  rep("03", 2),
  rep("04", 2)
)

check_rs_rsdtc_across_visit(RS1)
check_rs_rsdtc_across_visit(RS1, preproc = ql_derive_seq)

# example with RSTESTCD with records flagged
RS2 <- RS1
RS2$RSTESTCD <- c(
  rep("OVRLRESP", 2), rep("OTHER", 2),
  rep("OVRLRESP", 2), rep("OVRLRESP", 2), rep("OTHER", 2)
)
check_rs_rsdtc_across_visit(RS2)
check_rs_rsdtc_across_visit(RS2, preproc = ql_derive_seq)

# example with records flagged without xxSPID
RS3 <- RS
RS3$RSSPID <- NULL
check_rs_rsdtc_across_visit(RS3)
check_rs_rsdtc_across_visit(RS3, preproc = ql_derive_seq)

# example without required variable
RS4 <- RS
RS4$VISIT <- NULL
check_rs_rsdtc_across_visit(RS4)
check_rs_rsdtc_across_visit(RS4, preproc = ql_derive_seq)

}
\author{
Will Harris
}
