% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_check_ss_ssdtc_dead_ds.R
\name{check_ss_ssdtc_dead_ds}
\alias{check_ss_ssdtc_dead_ds}
\title{Check non-missing dead status date in SS and non-missing according DS record with death date
where status date is greater or equal to death date}
\usage{
check_ss_ssdtc_dead_ds(SS, DS, preproc = identity, ...)
}
\arguments{
\item{SS}{Subject Status SDTM dataset with variables USUBJID, SSDTC, SSSTRESC, VISIT, SSSPID (optional)}

\item{DS}{Disposition SDTM dataset with variables USUBJID, DSSTDTC, DSDECOD, DSCAT, DSSPID (optional)}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check looks for missing death date in DS dataset
if there is dead status date in SS dataset or
if Subject Status Date/Time of Assessments is less than
Start Date/Time of Disposition Event(SS.SSDTC < DS.DSSTDTC)
}
\examples{

SS <- data.frame(
  USUBJID = 1:5,
  SSDTC = "2020-01-02",
  SSSTRESC = c("死亡", "死亡", "存活", "死亡", "存活"),
  VISIT = "FOLLOW-UP",
  SSSEQ = 1:5,
  SSSPID = sprintf("\%02d", 1:5)
)

DS <- data.frame(
  USUBJID = 1:5,
  DSSTDTC = c("2020-01-02", "2020-01-02", "2020-01-01", "2020-01-03", "2020-01-01"),
  DSDECOD = c(rep("死亡", 5)),
  DSCAT = c("OTHER EVENT", rep("DISPOSITION EVENT", 4)),
  DSSEQ = 1:5,
  DSSPID = sprintf("\%02d", 1:5)
)

check_ss_ssdtc_dead_ds(SS, DS)
check_ss_ssdtc_dead_ds(SS, DS, preproc = ql_derive_seq)


SS <- data.frame(
  USUBJID = 1:5,
  SSDTC = "2020-01-02",
  SSSTRESC = c(rep("", 5)),
  VISIT = "FOLLOW-UP",
  SSSEQ = 1:5,
  SSSPID = sprintf("\%02d", 1:5)
)

DS <- data.frame(
  USUBJID = 1:5,
  DSSTDTC = c("2020-01-02", "2020-01-02", "2020-01-01", "2020-01-03", "2020-01-01"),
  DSDECOD = c(rep("死亡", 5)),
  DSCAT = c(rep("DISPOSITION EVENT", 5)),
  DSSEQ = 1:5,
  DSSPID = sprintf("\%02d", 1:5)
)

check_ss_ssdtc_dead_ds(SS, DS)
check_ss_ssdtc_dead_ds(SS, DS, preproc = ql_derive_seq)


SS <- data.frame(
  USUBJID = 1:5,
  SSDTC = "2020-01-02",
  SSSTRESC = c(rep("", 5)),
  VISIT = "FOLLOW-UP",
  SSSEQ = 1:5,
  SSSPID = sprintf("\%02d", 1:5)
)

DS <- data.frame(
  USUBJID = 1:5,
  DSSTDTC = 2,
  DSDECOD = c(rep("死亡", 5)),
  DSCAT = c(rep("DISPOSITION EVENT", 5)),
  DSSEQ = 1:5,
  DSSPID = sprintf("\%02d", 1:5)
)

check_ss_ssdtc_dead_ds(SS, DS)

}
\author{
Vira Vrakina
}
