% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ss_ssstat_ssorres.R
\name{check_ss_ssstat_ssorres}
\alias{check_ss_ssstat_ssorres}
\title{Check for non-missing SSORRES if SSSTAT is NOT DONE}
\usage{
check_ss_ssstat_ssorres(SS, preproc = identity, ...)
}
\arguments{
\item{SS}{Long-Term Survival Follow-Up SDTM dataset with variables USUBJID,
VISIT, SSSTAT, SSDTC, SSORRES, SSSPID (optional)}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check is for studies with LTFU mapped to the SS domain,
check that if '未查' (Unable to Contact), then there should not be
a response (SSORRES)
}
\examples{

SS <- data.frame(
  STUDYID = 1,
  USUBJID = c(rep(1, 6), rep(2, 6)),
  SSSTRESC = c("存活", "死亡", "存活", "", "", "未知", "存活", "死亡", "存活", "", "", "未知"),
  SSORRES = c("存活", "死亡", "存活", "", "", "未知", "存活", "死亡", "存活", "", "", "未知"),
  VISIT = rep(paste0("生存随访", 1:6), 2),
  SSSTAT = rep(c("", "未查"), 6),
  SSDTC = rep("2016-01-01", 12),
  SSSEQ = rep(1:6, 2),
  SSSPID = sprintf("\%02d", rep(1:6, 2)),
  stringsAsFactors = FALSE
)

check_ss_ssstat_ssorres(SS)

SS$SSORRES[2] <- NA
check_ss_ssstat_ssorres(SS)

SS$SSSPID <- "01"
check_ss_ssstat_ssorres(SS, preproc = ql_derive_seq)

SS$SSORRES[6] <- NA
SS$SSORRES[8] <- ""
SS$SSORRES[12] <- NA
check_ss_ssstat_ssorres(SS)

SS$SSORRES <- NULL
check_ss_ssstat_ssorres(SS)

}
\author{
1
}
