% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tu_tudtc.R
\name{check_tu_tudtc}
\alias{check_tu_tudtc}
\title{Check for missing TUDTC values}
\usage{
check_tu_tudtc(TU, preproc = identity, ...)
}
\arguments{
\item{TU}{Tumor Identification SDTM dataset with variables USUBJID, TUDTC,
VISIT, TUORRES, TUEVAL (optional), TUTESTCD (optional), TULNKID (optional), TUSPID (optional)}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check looks for missing TUDTC values and returns a data frame.
If TUEVAL exists, only records with TUEVAL = 'з ”з©¶иЂ…' are checked.
If TUEVAL does not exist, all records are checked.
}
\examples{
TU <- data.frame(
  USUBJID = "1001",
  TUDTC = "2020-05-05",
  VISIT = "C1D1",
  TUORRES = rep(c("йќ¶з—…зЃ¶", "йќћйќ¶з—…зЃ¶"), 5),
  TUSEQ = 1:10,
  TULNKID = sprintf("T\%02d", 1:10),
  TUSPID = sprintf("\%02d", 1:10),
  TUEVAL = "з ”з©¶иЂ…",
  TUTESTCD = "TUMIDENT",
  stringsAsFactors = FALSE
)

TU$TUDTC[1] <- ""
TU$TUDTC[2] <- "NA"
TU$TUDTC[3] <- NA

check_tu_tudtc(TU, preproc = ql_derive_seq)

TU$TUEVAL[1] <- ""
TU$TUTESTCD <- NULL
check_tu_tudtc(TU, preproc = ql_derive_seq)

TU$TUEVAL[2] <- "з‹¬з«‹иЇ„дј°иЂ…"
TU$TUEVAL[3] <- "з‹¬з«‹иЇ„дј°иЂ…"
TU$TUDTC[4] <- ""
check_tu_tudtc(TU)

TU$TUSPID <- NULL
check_tu_tudtc(TU)

TU$VISIT <- NULL
check_tu_tudtc(TU)

## PASS example
TU_PASS <- data.frame(
  USUBJID = c("1001", "1002"),
  TUORRES = c("йќ¶з—…зЃ¶", "йќћйќ¶з—…зЃ¶"),
  TUDTC = c("2020-05-05", "2020-05-06"),
  VISIT = c("C1D1", "C1D1"),
  TUORRES = c("A", "B"),
  TUEVAL = c("з ”з©¶иЂ…", "з‹¬з«‹иЇ„дј°иЂ…"),
  stringsAsFactors = FALSE
)
check_tu_tudtc(TU_PASS)

}
\author{
Beeya Na
}
