% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vs_vsdtc_after_dd.R
\name{check_vs_vsdtc_after_dd}
\alias{check_vs_vsdtc_after_dd}
\title{Check for VS dates occurring after death date}
\usage{
check_vs_vsdtc_after_dd(DM, VS)
}
\arguments{
\item{DM}{Demographics SDTM dataset with variables USUBJID,
DTHDTC}

\item{VS}{Vital Signs SDTM dataset with variables USUBJID,
VSDTC, VSTESTCD, and VSORRES}
}
\value{
Boolean value for whether the check passed or failed, with 'msg'
attribute if the check failed
}
\description{
This check looks for VS dates that occur after death date
}
\examples{

DM <- data.frame(
  STUDYID = 1:5, USUBJID = LETTERS[1:5],
  DTHDTC = c(rep("", 4), "2016-01-02"),
  stringsAsFactors = FALSE
)

VS <- data.frame(
  STUDYID = 1:5, USUBJID = LETTERS[1:5],
  VSDTC = rep("2015-12-31", 5),
  VSTESTCD = letters[1:5],
  VSORRES = 1:5,
  stringsAsFactors = FALSE
)

check_vs_vsdtc_after_dd(DM, VS)

VS$VSDTC[1] <- "2016-01-03"
VS$USUBJID[1] <- VS$USUBJID[5]

check_vs_vsdtc_after_dd(DM, VS)

}
\author{
1
}
