% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtmchecks-package.R
\docType{package}
\name{Emei-package}
\alias{Emei}
\alias{Emei-package}
\title{Emei: EMEI-SDTM Edit-rule Monitoring & Error Inspection}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The purpose of the Emei package is to help detect and investigate potential analysis relevant issues in SDTM data. This is done using a set of data check functions. These check functions are intended to be generalizable, actionable, and meaningful for analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/r.pkg.template/}
  \item Report bugs at \url{https://github.com/insightsengineering/r.pkg.template/issues}
}

}
\author{
\strong{Maintainer}: insightsengineering \email{insightsengineering@example.com}

}
\keyword{internal}
