% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lb_lbdtc_after_dd.R
\name{check_lb_lbdtc_after_dd}
\alias{check_lb_lbdtc_after_dd}
\title{Check for LB dates occurring after death date}
\usage{
check_lb_lbdtc_after_dd(DM, LB)
}
\arguments{
\item{DM}{Demographics SDTM dataset with variables USUBJID,
DTHDTC}

\item{LB}{Laboratory Test Findings SDTM dataset with variables USUBJID,
LBDTC, LBTESTCD, and LBORRES. LBSTAT is optional.
LBTESTCD is included in output for context.}
}
\value{
Boolean value for whether the check passed or failed, with 'msg'
attribute if the check failed
}
\description{
This check looks for LB dates that occur after death date
}
\examples{

DM <- data.frame(
  STUDYID = 1:5, USUBJID = LETTERS[1:5],
  DTHDTC = c(rep("", 4), "2016-01-02"),
  stringsAsFactors = FALSE
)

LB <- data.frame(
  STUDYID = 1:5, USUBJID = LETTERS[1:5],
  LBDTC = rep("2015-12-31", 5),
  LBTESTCD = letters[1:5],
  LBORRES = 1:5,
  LBSTAT = "",
  stringsAsFactors = FALSE
)

check_lb_lbdtc_after_dd(DM, LB)

LB$LBDTC[1] <- "2016-01-03"
LB$USUBJID[1] <- LB$USUBJID[5]

check_lb_lbdtc_after_dd(DM, LB)

LB$LBSTAT[1] <- "未查"
check_lb_lbdtc_after_dd(DM, LB)

}
\author{
1
}
