% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ss_ssdtc_alive_dm.R
\name{check_ss_ssdtc_alive_dm}
\alias{check_ss_ssdtc_alive_dm}
\title{Check SS records with alive result and non-missing SSDTC later than DM death date}
\usage{
check_ss_ssdtc_alive_dm(SS, DM, preproc = identity, ...)
}
\arguments{
\item{SS}{Subject Status SDTM dataset with variables USUBJID, SSDTC, SSORRES, SSTESTCD, VISIT, SSSPID (optional), SEQ (optional)}

\item{DM}{Demographics SDTM dataset with variables USUBJID, DTHDTC}

\item{preproc}{An optional company specific preprocessing script}

\item{...}{Other arguments passed to methods}
}
\value{
boolean value if check failed or passed with 'msg' attribute if the
test failed
}
\description{
This check looks for non-missing SS.SSDTC when SS.SSTESTCD='SURVSTAT' and
SS.SSORRES='存活', then compares SS.SSDTC with DM.DTHDTC
by subject and reports records where SS.SSDTC > DM.DTHDTC.
}
\examples{

SS <- data.frame(
  USUBJID = 1:5,
  SSSEQ = 1:5,
  SSDTC = "2020-01-02",
  SSTESTCD = "SURVSTAT",
  SSORRES = c("死亡", "死亡", "存活", "死亡", "存活"),
  VISIT = "生存随访"
)

DM <- data.frame(
  USUBJID = 1:5,
  DTHDTC = "2020-01-03"
)

check_ss_ssdtc_alive_dm(SS, DM)
check_ss_ssdtc_alive_dm(SS, DM, preproc = ql_derive_seq)

SS1 <- data.frame(
  USUBJID = 1:5,
  SSSEQ = 1:5,
  SSDTC = "2020-01-04",
  SSTESTCD = "SURVSTAT",
  SSORRES = c("死亡", "死亡", "存活", "死亡", "存活"),
  VISIT = "生存随访"
)

DM <- data.frame(
  USUBJID = 1:5,
  DTHDTC = c("2020-01-04", "2020-01-05", "2020-01-03", "2020-01-04", "2020-01-05")
)

check_ss_ssdtc_alive_dm(SS1, DM)
check_ss_ssdtc_alive_dm(SS1, DM, preproc = ql_derive_seq)

}
\author{
1
}
